feat: Refactor is_prime module and route
- Refactor the is_prime module and route to use the generate_random_prime module for prime number generation.
- Removed the is_prime.rs file as it is no longer needed.
- Updated the main.rs file to use the generate_random_prime module for prime number checking.
