+++ src/generate_random_prime.rs
@@ -1,22 +1,31 @@
 use rand::Rng;
 
-fn is_prime(n: u64) -> bool {
-    if n <= 1 {
+pub fn is_prime(num: u64) -> bool {
+    if num <= 1 {
         return false;
     }
-    if n <= 3 {
+    if num <= 3 {
         return true;
     }
-    if n % 2 == 0 || n % 3 == 0 {
+    if num % 2 == 0 || num % 3 == 0 {
         return false;
     }
-    let mut i = 5;
-    while i * i <= n {
-        if n % i == 0 || n % (i + 2) == 0 {
+
+    let sqrt_num = (num as f64).sqrt().floor() as usize;
+    for i in (5..=sqrt_num).step_by(6) {
+        if num % (i as u64) == 0 || num % ((i + 2) as u64) == 0 {
             return false;
         }
-        i += 6;
     }
+
+    let mut i = 3;
+    while i * i <= num {
+        if num % i == 0 {
+            return false;
+        }
+        i += 2;
+    }
+
     true
 }
